RolePermission test & policies

+++ modules/abouterf/Course/Http/Controllers/CourseController.php
@@ -4,6 +4,7 @@ namespace abouterf\Course\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use abouterf\Category\Repositories\CategoryRepo;
+use abouterf\Category\Responses\AjaxResponse;
 use abouterf\Category\Responses\AjaxResponses;
 use abouterf\Course\Http\Requests\CourseRequest;
 use abouterf\Course\Models\Course;
@@ -47,9 +48,12 @@ class CourseController extends Controller
     public function update($id, CourseRequest $request, CourseRepo $courseRepo)
     {
         $course = $courseRepo->findByid($id);
+        $this->authorize('edit', $course);
         if ($request->hasFile('image')) {
             $request->request->add(['banner_id' => MediaFileService::upload($request->file('image'))->id]);
-            $course->banner->delete();
+            if ($course->banner) {
+                $course->banner->delete();
+            }
         } else {
             $request->request->add(['banner_id' => $course->banner_id]);
         }
@@ -60,6 +64,7 @@ class CourseController extends Controller
     public function destroy($id, CourseRepo $courseRepo)
     {
         $course =  $courseRepo->findByid($id);
+        $this->authorize('destroy', $course);
 
         if ($course->banner) {
             $course->banner->delete();
@@ -67,33 +72,39 @@ class CourseController extends Controller
 
         $course->delete();
 
-        return AjaxResponses::SuccessResponse();
+        return AjaxResponse::SuccessResponse();
     }
 
     public function accept($id, CourseRepo $courseRepo)
     {
+        $this->authorize('change_confirmation_status', Course::class);
+
         if ($courseRepo->updateConfirmationStatus($id, Course::CONFIRMATION_STATUS_ACCEPTED)) {
-            return AjaxResponses::SuccessResponse();
+            return AjaxResponse::SuccessResponse();
         }
 
-        return AjaxResponses::FailedResponse();
+        return AjaxResponse::FailedResponse();
     }
 
     public function reject($id, CourseRepo $courseRepo)
     {
+        $this->authorize('change_confirmation_status', Course::class);
+
         if ($courseRepo->updateConfirmationStatus($id, Course::CONFIRMATION_STATUS_REJECTED)) {
-            return AjaxResponses::SuccessResponse();
+            return AjaxResponse::SuccessResponse();
         }
 
-        return AjaxResponses::FailedResponse();
+        return AjaxResponse::FailedResponse();
     }
 
     public function lock($id, CourseRepo $courseRepo)
     {
+        $this->authorize('change_confirmation_status', Course::class);
+
         if ($courseRepo->updateStatus($id, Course::STATUS_LOCKED)) {
-            return AjaxResponses::SuccessResponse();
+            return AjaxResponse::SuccessResponse();
         }
 
-        return AjaxResponses::FailedResponse();
+        return AjaxResponse::FailedResponse();
     }
 }
